requirements.txt
asgiref==3.3.1
astroid==2.5.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
defusedxml==0.6.0
Django==3.1.7
django-environ==0.4.5
django-filter==2.4.0
django-model-utils==4.1.1
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
isort==5.7.0
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.9
Pillow==8.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
pylint==2.7.2
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
wrapt==1.12.1
